Arrange tests and kafka standalon
