+++ examples/spring-example/src/test/kotlin/com/stove/spring/example/e2e/TestSystemInitializer.kt
