--- lib/stove-testing-e2e-kafka/src/main/kotlin/com/trendyol/stove/testing/e2e/kafka/intercepting/InterceptionOptions.kt
@@ -1,8 +1,0 @@
-package com.trendyol.stove.testing.e2e.kafka.intercepting
-
-class InterceptionOptions(
-    private val errorTopicSuffixes: List<String>,
-) {
-
-    fun isErrorTopic(topic: String): Boolean = errorTopicSuffixes.any { suffix -> topic.endsWith(suffix, ignoreCase = true) }
-}
