--- lib/stove-testing-e2e-kafka/src/main/kotlin/com/trendyol/stove/testing/e2e/kafka/intercepting/RecordsAssertions.kt
@@ -1,8 +1,0 @@
-package com.trendyol.stove.testing.e2e.kafka.intercepting
-
-import java.util.UUID
-import java.util.concurrent.ConcurrentMap
-
-internal interface RecordsAssertions {
-    val assertions: ConcurrentMap<UUID, KafkaAssertion<*>>
-}
