+++ lib/stove-testing-e2e-kafka/src/main/kotlin/com/trendyol/stove/testing/e2e/standalone/kafka/intercepting/Failure.kt
@@ -1,0 +1,7 @@
+package com.trendyol.stove.testing.e2e.standalone.kafka.intercepting
+
+data class Failure(
+    val topic: String,
+    val message: Any,
+    val reason: Throwable,
+)
