+++ lib/stove-testing-e2e-kafka/src/main/kotlin/com/trendyol/stove/testing/e2e/standalone/kafka/intercepting/RecordsAssertions.kt
@@ -1,0 +1,8 @@
+package com.trendyol.stove.testing.e2e.standalone.kafka.intercepting
+
+import java.util.UUID
+import java.util.concurrent.ConcurrentMap
+
+internal interface RecordsAssertions {
+    val assertions: ConcurrentMap<UUID, KafkaAssertion<*>>
+}
