+++ lib/stove-testing-e2e-kafka/src/test/kotlin/com/trendyol/stove/testing/e2e/standalone/kafka/setup/example/consumers/ProductConsumer.kt
@@ -1,0 +1,15 @@
+package com.trendyol.stove.testing.e2e.standalone.kafka.setup.example.consumers
+
+import com.trendyol.stove.testing.e2e.standalone.kafka.setup.example.KafkaTestShared.TopicDefinition
+import com.trendyol.stove.testing.e2e.standalone.kafka.setup.example.StoveListener
+import org.apache.kafka.clients.consumer.ConsumerRecord
+
+class ProductConsumer(
+    consumerSettings: Map<String, Any>,
+    producerSettings: Map<String, Any>,
+) : StoveListener(consumerSettings, producerSettings) {
+
+    override val topicDefinition: TopicDefinition = TopicDefinition("product", "product.retry", "product.error")
+    override suspend fun listen(record: ConsumerRecord<String, String>) {
+    }
+}
